add custom_action to custom file
