+++ lux/action/custom.py
@@ -2,6 +2,7 @@ from lux.interestingness.interestingness import interestingness
 import lux
 from lux.executor.PandasExecutor import PandasExecutor
 from lux.executor.SQLExecutor import SQLExecutor
+from lux._config.config import actions
 #for benchmarking
 import time
 
@@ -31,3 +32,17 @@ def custom(ldf):
     # ldf.clear_intent()
     vlist.sort(remove_invalid=True)
     return recommendation
+
+def custom_action(ldf):
+    if (actions.__len__() > 0):
+        for action_name in actions.__dir__():
+            validator = actions.__getattr__(action_name).validator(ldf)
+            if validator:
+                function = actions.__getattr__(action_name).function(ldf)
+                recommendation = {"action":function["action"], "description":function["description"]}
+                recommendation["collection"] = function["collection"]
+                return recommendation
+            else:
+                return None
+    else:
+        return None
